refactor: split photo service responsibilities
