+++ backend/PhotoBank.DependencyInjection/AddPhotobankCoreExtensions.cs
@@ -6,6 +6,9 @@ using PhotoBank.Repositories;
 using PhotoBank.Services;
 using PhotoBank.Services.Api;
 using PhotoBank.Services.Internal;
+using PhotoBank.Services.Photos.Admin;
+using PhotoBank.Services.Photos.Faces;
+using PhotoBank.Services.Photos.Queries;
 using PhotoBank.Services.Search;
 using PhotoBank.Services.Translator;
 using Polly;
@@ -31,6 +34,11 @@ public static partial class ServiceCollectionExtensions
         services.AddScoped<IFaceStorageService, FaceStorageService>();
         services.AddScoped<MinioObjectService>();
         services.AddScoped<ISearchFilterNormalizer, SearchFilterNormalizer>();
+        services.AddScoped<IPhotoQueryService, PhotoQueryService>();
+        services.AddScoped<IPersonDirectoryService, PersonDirectoryService>();
+        services.AddScoped<IPersonGroupService, PersonGroupService>();
+        services.AddScoped<IFaceCatalogService, FaceCatalogService>();
+        services.AddScoped<IPhotoAdminService, PhotoAdminService>();
         services.AddScoped<IPhotoService, PhotoService>();
         services.AddScoped<ISearchReferenceDataService, SearchReferenceDataService>();
         services.AddPhotoEvents();
